chore: extract node compat helper to own file
