+++ packages/wrangler/src/type-generation/index.ts
@@ -4,11 +4,11 @@ import * as esmLexer from "es-module-lexer";
 import { findUpSync } from "find-up";
 import { findWranglerToml, readConfig } from "../config";
 import { getEntry } from "../deployment-bundle/entry";
-import { getNodeCompatMode } from "../deployment-bundle/node-compat";
 import { getVarsForDev } from "../dev/dev-vars";
 import { UserError } from "../errors";
 import { CommandLineArgsError } from "../index";
 import { logger } from "../logger";
+import { getNodeCompatMode } from "../node-compat";
 import { parseJSONC } from "../parse";
 import { printWranglerBanner } from "../update-check";
 import { generateRuntimeTypes } from "./runtime";
